refactor: addressed comments and change route to routeGen4 and routeLegacy back to route (v3)
