+++ go/vt/vtgate/planbuilder/filtering.go
@@ -53,7 +53,7 @@ func planFilter(pb *primitiveBuilder, input logicalPlan, filter sqlparser.Expr, 
 		}
 		return node, nil
 
-	case *routeLegacy:
+	case *route:
 		sel := node.Select.(*sqlparser.Select)
 		switch whereType {
 		case sqlparser.WhereStr:
