+++ go/vt/vtgate/planbuilder/ordering.go
@@ -49,7 +49,7 @@ func planOrdering(pb *primitiveBuilder, input logicalPlan, orderBy v3OrderBy) (l
 		}
 		node.underlying = plan
 		return node, nil
-	case *routeLegacy:
+	case *route:
 		return planRouteOrdering(orderBy, node)
 	case *join:
 		return planJoinOrdering(pb, orderBy, node)
@@ -243,7 +243,7 @@ func planJoinOrdering(pb *primitiveBuilder, orderBy v3OrderBy, node *join) (logi
 	return node, nil
 }
 
-func planRouteOrdering(orderBy v3OrderBy, node *routeLegacy) (logicalPlan, error) {
+func planRouteOrdering(orderBy v3OrderBy, node *route) (logicalPlan, error) {
 	switch len(orderBy) {
 	case 0:
 		return node, nil
