+++ go/vt/vtgate/planbuilder/testdata/ddl_cases.txt
@@ -264,7 +264,7 @@ Gen4 plan same as above
     "Query": "create view view_a as select id, `name` from unsharded where id in (select id from unsharded where id = 1 union select id from unsharded where id = 3)"
   }
 }
-Gen4 error: expr cannot be converted, not supported: sqlparser.ListArg
+Gen4 plan same as above
 
 # create view with subquery in unsharded keyspace with UNION clause
 "create view view_a as (select id from unsharded) union (select id from unsharded_auto) order by id limit 5"
